moved data directory to be in test
