+++ coco/test/oldconfig/config.go
@@ -1,4 +1,4 @@
-package coco
+package oldconfig
 
 import (
 	"bytes"
