+++ coco/test/oldconfig/config_test.go
@@ -1,4 +1,4 @@
-package coco
+package oldconfig
 
 import (
 	"sync"
