+++ coco/test/oldconfig/graphs_test.go
@@ -1,4 +1,4 @@
-package coco
+package oldconfig
 
 import (
 	"fmt"
